S3A.9b: relocate emit_bg_histogram_kernel out of dialects/parallel/data
Per docs/stage3a_execution_plan.md §7 task S3A.9b.

Before: dialects/parallel/data/block_group.py contained both:
  - The dialect's data definitions (BgRootCjMulti, BgSourceSpec).
  - A 144-line CUDA emission function `emit_bg_histogram_kernel`
    that pulls in codegen/cuda/{context,plugin,view_slots}.

The emission function violated the layering invariant from
ir/core/CLAUDE.md: a dialect should not contain target rendering.

After: emit_bg_histogram_kernel lives in
codegen/cuda/render/parallel_data.py alongside the BgRootCjMulti
renderer. The dialect file now contains only ops + helper
dataclasses. block_group.py no longer imports srdatalog.ir.mir.types.

Why this kernel template lives in the codegen (not as a
@register_render handler): it's a per-rule kernel signature
generator called by complete_runner.py during runner emit, not
a body-shaped IIR renderer dispatched per-op. Both kinds of
emission legitimately live in the codegen — one routes through
the @register_render registry (per-op IIR renderers), the other
is a top-level kernel template called directly.

Updated 1 import site in codegen/cuda/complete_runner.py + 1
docstring reference in dialects/relation/d2l/__init__.py.

Test gate: 1057 passed, 5 skipped — baseline preserved.

Bundle A commit 5/5. Bundle complete:
  A.1 — MIR rename (mir/emit.py → mir/print.py)
  A.2 — sorted_array lowerings no longer call codegen at lowering time
  A.3 (gitignore .claude/)
  A.4 — split codegen/cuda/emit.py into render/ package
  A.5 — relocate emit_bg_histogram_kernel (this commit)
